requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.8
cycler==0.11.0
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lxml==4.6.4
Markdown==3.3.6
matplotlib==3.5.0
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
pandas-datareader==0.10.0
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
sklearn==0.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tomli==1.2.2
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
absl-py==0.13.0
appdirs==1.4.4
astunparse==1.6.3
bcrypt==3.2.0
CacheControl==0.12.6
cachetools==4.2.2
cachy==0.3.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
cleo==0.8.1
clikit==0.6.2
colorama==0.4.4
crashtest==0.3.1
decorator==5.0.9
distlib==0.3.2
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
html5lib==1.1
idna==2.10
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==21.8.0
lockfile==0.12.2
Markdown==3.3.4
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.7.1
poetry==1.1.9
poetry-core==1.0.5
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.9.0
pylev==1.4.0
pyparsing==2.4.7
pyte==0.8.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
shellingham==1.4.0
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
thefuck==3.31
tomlkit==0.7.2
typing-extensions==3.7.4.3
urllib3==1.26.5
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
absl-py==0.13.0
appdirs==1.4.4
astunparse==1.6.3
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
bcrypt==3.2.0
binance-connector==1.8.0
CacheControl==0.12.6
cachetools==4.2.2
cachy==0.3.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
cleo==0.8.1
clikit==0.6.2
colorama==0.4.4
constantly==15.1.0
crashtest==0.3.1
cryptography==36.0.0
decorator==5.0.9
distlib==0.3.2
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
html5lib==1.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==21.8.0
lockfile==0.12.2
Markdown==3.3.4
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pkginfo==1.7.1
poetry==1.1.9
poetry-core==1.0.5
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.9.0
pylev==1.4.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyte==0.8.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.7.2
service-identity==21.1.0
shellingham==1.4.0
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
thefuck==3.31
tomlkit==0.7.2
Twisted==21.2.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.5
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
zope.interface==5.4.0
